Added new required flag to TargetParameterItem + improved processor type usage
This commit also improves code quality and removes the dependency on the archtree library. It refactores the AutoTargetFragmentDispatcher class and improves the way the processor uses types (provided by the annotations). It fixes the bug that it was not possible to define MyParcelable (for a target) but always had to be Parcelable.