Refactored the generator for bundle-models and target-methods
For bundle-model: The bundle used for the model always needs to non-null. That prevents misusage of the relevant model.

For target-methods: There are now two types of method - required only and required+optional. Each TargetParameterItem got a new constructor-paramter that must be defined if the item is required or not. Non-required parameters will be added to lists or to a optional default target method. This commit also changes the usage of non-null/nullable annotation. Required parameter (non-primitive) are always non-null. That should make the usage with kotlin much easier.

This commit also adds documentation to the different annotation classes.