Tweaking TargetService and added more kotlin doc
This commit removes any dagger dependencies from the lib core artifact. The TargetService won't be added to the dagger dependency graph automatically and requires a Provide based method. Additionally the TargetService constructor requires a ContextProviderCommunicator instead of the implementation. This should simplify unit tests that are trying to mock the TargetService object.